add trait for row to string converting without file
